+++ activerecord/test/cases/autosave_association_test.rb
@@ -1951,3 +1951,75 @@ class TestAutosaveAssociationOnAHasManyAssociationDefinedInSubclassWithAcceptsNe
     assert_equal "Updated", valid_project.name
   end
 end
+
+class TestCyclicAutosaveAssociationsOnlySaveOnce < ActiveRecord::TestCase
+  teardown do
+    $autosave_saving_stack = []
+  end
+
+  test "child is saved only once if child is the inverse has_one of parent" do
+    ship_reflection = Ship.reflect_on_association(:pirate)
+    pirate_reflection = Pirate.reflect_on_association(:ship)
+    assert_equal ship_reflection, pirate_reflection.inverse_of, "The pirate reflection's inverse should be the ship reflection"
+
+    child = Ship.new(name: "Nights Dirty Lightning")
+    parent = child.build_pirate(catchphrase: "Aye")
+    child.save!
+    assert child.previously_new_record?
+    assert parent.previously_new_record?
+  end
+
+  test "child is saved only once if child is an inverse has_many of parent" do
+    child = FamousShip.new(name: "Poison Orchid")
+    parent = child.build_famous_pirate(catchphrase: "Aye")
+    child.save!
+    assert child.previously_new_record?
+    assert parent.previously_new_record?
+  end
+
+  test "similar children are saved in the autosave" do
+    child1 = FamousShip.new(name: "Poison Orchid")
+    parent = child1.build_famous_pirate(catchphrase: "Aye")
+    child2 = parent.famous_ships.build(name: "Red Messenger")
+    child1.save!
+    assert child2.persisted?
+    assert child1.previously_new_record?
+    assert parent.previously_new_record?
+    assert_equal [child2, child1], parent.reload.famous_ships
+  end
+
+  test "parent is saved only once" do
+    child = Ship.new(name: "Nights Dirty Lightning")
+    parent = child.build_pirate(catchphrase: "Aye")
+    parent.save!
+    assert child.previously_new_record?
+    assert parent.previously_new_record?
+  end
+
+  test "saving? is reset to false if validations fail" do
+    child = Ship.new(name: "Nights Dirty Lightning")
+    child.build_pirate
+    assert_not child.save
+    assert_not_predicate child, :saving?
+  end
+
+  test "saving? is set to false after multiple nested saves" do
+    ship_with_saving_stack = Class.new(Ship) do
+      before_save { $autosave_saving_stack << saving? }
+      after_save  { $autosave_saving_stack << saving? }
+    end
+
+    pirate_with_callbacks = Class.new(Pirate) do
+      after_save   { ship.save }
+      after_create { ship.save }
+      after_commit { ship.save }
+    end
+
+    child = ship_with_saving_stack.new(name: "Nights Dirty Lightning")
+    child.pirate = pirate_with_callbacks.new(catchphrase: "Aye")
+    $autosave_saving_stack = []
+    child.save!
+    assert_equal [true] * 8, $autosave_saving_stack
+    assert_equal false, child.saving?
+  end
+end