+++ activerecord/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Prevent double saves in autosave of cyclic associations
+
+    Adds a @saving state which tracks if a record is currently being saved.
+    If @saving is set to true, the record won't be saved by the autosave callbacks.
+
+    *Petrik de Heus*
+
 *   Fix Float::INFINITY assignment to datetime column with postgresql adapter
 
     Before: